requirements.txt
torch>=2.0.0
torchvision
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
safetensors>=0.4.0
controlnet-aux>=0.0.7
opencv-python>=4.8.0
Pillow>=10.0.0
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
python-socketio>=5.10.0
click>=8.1.0
tqdm>=4.66.0
numpy>=1.24.0
scikit-image>=0.21.0
